refactor: router navigate and run logic
