+++ lib/routers/index.js
@@ -1,1 +1,1 @@
-export * from "./router";
+export * from "./create-router";
